Trying to solve the integer overflow in davidson
I got an integer overflow, due to basis_size * max_vecs.
Trying to fix it. changing basis_size to long long should make the
multiplcation of basis_size * max_vecs also happen at long long size. If
this does not fix it, size_t that is passed to malloc is probably not
large enough.

+++ include/wrapper_solvers.h
@@ -37,7 +37,7 @@
  * \param [in] davidson_max_vec Number of maximum vectors to be taken into
  * account before deflation is needed in the Davidson algorithm.
  */
-int sparse_eigensolve(double * vec, int dims, double* energy, 
+int sparse_eigensolve(double * vec, long long dims, double* energy, 
                       void (*matvec)(double*, double*, void*), 
                       const double * diagonal, double tol, int max_its, 
                       const char solver[], int davidson_keep,
